rastertek - Bump Mapping

+++ rastertek/Input.cpp
@@ -113,12 +113,12 @@ namespace wf
 
 	bool Input::IsEscapePressed()
 	{
-		return m_keyboard_state[ DIK_ESCAPE ] & 0x80;
+		return static_cast<int>( m_keyboard_state[ DIK_ESCAPE ] ) & 0x80;
 	}
 
 	bool Input::IsSpacePressed()
 	{
-		return m_keyboard_state[ DIK_SPACE ] & 0x80;
+		return static_cast<int>( m_keyboard_state[ DIK_SPACE ] ) & 0x80;
 	}
 
 	void Input::GetMouseLocation( int& _x, int& _y )
